backend/requirements.txt
fastapi
uvicorn
openai-whisper
ffmpeg-python
torch
pydub